PEPPER-1402 Moved DsmNotificationEventType.java and DsmNotificationPayload.java and KitReasonType.java into ddp-common to be accessible through both dss-core and dsm-core
