+++ pepper-apis/ddp-common/src/main/java/org/broadinstitute/ddp/notficationevent/DsmNotificationEventType.java
@@ -1,4 +1,4 @@
-package org.broadinstitute.ddp.model.dsm;
+package org.broadinstitute.ddp.notficationevent;
 
 public enum DsmNotificationEventType {
     /**
