+++ pepper-apis/dss-core/src/main/java/org/broadinstitute/ddp/db/dao/EventTriggerDao.java
@@ -6,8 +6,8 @@ import org.broadinstitute.ddp.db.DBUtils;
 import org.broadinstitute.ddp.db.DaoException;
 import org.broadinstitute.ddp.model.activity.types.EventTriggerType;
 import org.broadinstitute.ddp.model.activity.types.InstanceStatusType;
-import org.broadinstitute.ddp.model.dsm.DsmNotificationEventType;
 import org.broadinstitute.ddp.model.user.EnrollmentStatusType;
+import org.broadinstitute.ddp.notficationevent.DsmNotificationEventType;
 import org.jdbi.v3.sqlobject.CreateSqlObject;
 
 public interface EventTriggerDao {
