+++ pepper-apis/dss-core/src/main/java/org/broadinstitute/ddp/db/dao/EventTriggerSql.java
@@ -1,7 +1,7 @@
 package org.broadinstitute.ddp.db.dao;
 
 import org.broadinstitute.ddp.model.activity.types.EventTriggerType;
-import org.broadinstitute.ddp.model.dsm.DsmNotificationEventType;
+import org.broadinstitute.ddp.notficationevent.DsmNotificationEventType;
 import org.broadinstitute.ddp.model.user.EnrollmentStatusType;
 import org.jdbi.v3.sqlobject.SqlObject;
 import org.jdbi.v3.sqlobject.customizer.Bind;
