+++ pepper-apis/dss-core/src/main/java/org/broadinstitute/ddp/route/ReceiveDsmNotificationRoute.java
@@ -17,10 +17,10 @@ import org.broadinstitute.ddp.db.dao.KitTypeDao;
 import org.broadinstitute.ddp.db.dto.StudyDto;
 import org.broadinstitute.ddp.db.dto.kit.KitConfigurationDto;
 import org.broadinstitute.ddp.exception.DDPException;
-import org.broadinstitute.ddp.json.dsm.DsmNotificationPayload;
+import org.broadinstitute.ddp.notficationevent.DsmNotificationPayload;
 import org.broadinstitute.ddp.json.errors.ApiError;
-import org.broadinstitute.ddp.model.dsm.DsmNotificationEventType;
-import org.broadinstitute.ddp.model.dsm.KitReasonType;
+import org.broadinstitute.ddp.notficationevent.DsmNotificationEventType;
+import org.broadinstitute.ddp.notficationevent.KitReasonType;
 import org.broadinstitute.ddp.model.dsm.KitType;
 import org.broadinstitute.ddp.model.dsm.TestResult;
 import org.broadinstitute.ddp.model.event.DsmNotificationSignal;
