+++ pepper-apis/dss-core/src/test/java/org/broadinstitute/ddp/pex/TreeWalkInterpreterTest.java
@@ -69,8 +69,8 @@ import org.broadinstitute.ddp.model.activity.types.PicklistSelectMode;
 import org.broadinstitute.ddp.model.activity.types.MatrixSelectMode;
 import org.broadinstitute.ddp.model.activity.types.TemplateType;
 import org.broadinstitute.ddp.model.activity.types.TextInputType;
-import org.broadinstitute.ddp.model.dsm.DsmNotificationEventType;
-import org.broadinstitute.ddp.model.dsm.KitReasonType;
+import org.broadinstitute.ddp.notficationevent.DsmNotificationEventType;
+import org.broadinstitute.ddp.notficationevent.KitReasonType;
 import org.broadinstitute.ddp.model.dsm.TestResult;
 import org.broadinstitute.ddp.model.event.ActivityInstanceStatusChangeSignal;
 import org.broadinstitute.ddp.model.event.DsmNotificationSignal;
