+++ pepper-apis/dss-core/src/test/java/org/broadinstitute/ddp/route/ReceiveDsmNotificationRouteTest.java
@@ -1,10 +1,10 @@
 package org.broadinstitute.ddp.route;
 
 import static io.restassured.RestAssured.given;
-import static org.broadinstitute.ddp.model.dsm.DsmNotificationEventType.SALIVA_RECEIVED;
-import static org.broadinstitute.ddp.model.dsm.DsmNotificationEventType.TESTBOSTON_RECEIVED;
-import static org.broadinstitute.ddp.model.dsm.DsmNotificationEventType.TESTBOSTON_SENT;
-import static org.broadinstitute.ddp.model.dsm.DsmNotificationEventType.TEST_RESULT;
+import static org.broadinstitute.ddp.notficationevent.DsmNotificationEventType.SALIVA_RECEIVED;
+import static org.broadinstitute.ddp.notficationevent.DsmNotificationEventType.TESTBOSTON_RECEIVED;
+import static org.broadinstitute.ddp.notficationevent.DsmNotificationEventType.TESTBOSTON_SENT;
+import static org.broadinstitute.ddp.notficationevent.DsmNotificationEventType.TEST_RESULT;
 import static org.hamcrest.Matchers.containsString;
 import static org.hamcrest.Matchers.equalTo;
 import static org.junit.Assert.assertEquals;
@@ -47,13 +47,13 @@ import org.broadinstitute.ddp.db.dto.QueuedEventDto;
 import org.broadinstitute.ddp.db.dto.QueuedNotificationDto;
 import org.broadinstitute.ddp.db.dto.SendgridEmailEventActionDto;
 import org.broadinstitute.ddp.db.dto.StudyDto;
-import org.broadinstitute.ddp.json.dsm.DsmNotificationPayload;
+import org.broadinstitute.ddp.notficationevent.DsmNotificationPayload;
 import org.broadinstitute.ddp.model.activity.definition.FormActivityDef;
 import org.broadinstitute.ddp.model.activity.definition.i18n.Translation;
 import org.broadinstitute.ddp.model.activity.instance.ActivityResponse;
 import org.broadinstitute.ddp.model.activity.revision.RevisionMetadata;
-import org.broadinstitute.ddp.model.dsm.DsmNotificationEventType;
-import org.broadinstitute.ddp.model.dsm.KitReasonType;
+import org.broadinstitute.ddp.notficationevent.DsmNotificationEventType;
+import org.broadinstitute.ddp.notficationevent.KitReasonType;
 import org.broadinstitute.ddp.model.dsm.TestResult;
 import org.broadinstitute.ddp.model.pex.Expression;
 import org.broadinstitute.ddp.model.user.EnrollmentStatusType;
